Remove message fetching from services that are sending messages to have the same functionality in all services. Fetching the messages should not be done by the services to avoid fetching multiple messages.
