+++ agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/CloudOffboardingService.java
@@ -1,0 +1,6 @@
+package com.dke.data.agrirouter.api.service.messaging;
+
+import com.dke.data.agrirouter.api.service.parameters.CloudOffboardingParameters;
+
+public interface CloudOffboardingService extends MessagingService<CloudOffboardingParameters> {
+}
