--- agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/cloud/OffboardingService.java
@@ -1,7 +1,0 @@
-package com.dke.data.agrirouter.api.service.messaging.cloud;
-
-import com.dke.data.agrirouter.api.service.parameters.CloudOffboardingParameters;
-
-public interface OffboardingService {
-  void offboard(CloudOffboardingParameters parameters);
-}
