+++ agrirouter-api-java-impl/src/main/java/com/dke/data/agrirouter/impl/messaging/rest/CloudOnboardingServiceImpl.java
@@ -1,7 +1,4 @@
-package com.dke.data.agrirouter.impl.messaging.rest.cloud;
-
-import static com.dke.data.agrirouter.impl.messaging.rest.MessageFetcher.DEFAULT_INTERVAL;
-import static com.dke.data.agrirouter.impl.messaging.rest.MessageFetcher.MAX_TRIES_BEFORE_FAILURE;
+package com.dke.data.agrirouter.impl.messaging.rest;
 
 import agrirouter.cloud.registration.CloudVirtualizedAppRegistration;
 import agrirouter.commons.MessageOuterClass;
@@ -11,33 +8,35 @@ import com.dke.data.agrirouter.api.dto.encoding.EncodedMessage;
 import com.dke.data.agrirouter.api.dto.messaging.FetchMessageResponse;
 import com.dke.data.agrirouter.api.dto.onboard.OnboardingResponse;
 import com.dke.data.agrirouter.api.exception.CouldNotOnboardVirtualCommunicationUnitException;
+import com.dke.data.agrirouter.api.service.messaging.CloudOnboardingService;
 import com.dke.data.agrirouter.api.service.messaging.FetchMessageService;
 import com.dke.data.agrirouter.api.service.messaging.encoding.DecodeMessageService;
 import com.dke.data.agrirouter.api.service.messaging.encoding.EncodeMessageService;
-import com.dke.data.agrirouter.api.service.messaging.cloud.OnboardingService;
 import com.dke.data.agrirouter.api.service.parameters.CloudOnboardingParameters;
 import com.dke.data.agrirouter.api.service.parameters.SendMessageParameters;
 import com.dke.data.agrirouter.impl.messaging.MessageEncoder;
 import com.dke.data.agrirouter.impl.messaging.encoding.DecodeMessageServiceImpl;
 import com.dke.data.agrirouter.impl.messaging.encoding.EncodeMessageServiceImpl;
-import com.dke.data.agrirouter.impl.messaging.rest.FetchMessageServiceImpl;
-import com.dke.data.agrirouter.impl.messaging.rest.MessageSender;
 import com.dke.data.agrirouter.impl.validation.ResponseValidator;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-public class OnboardingServiceImpl
-    implements OnboardingService, MessageSender, ResponseValidator, MessageEncoder {
+import static com.dke.data.agrirouter.impl.messaging.rest.MessageFetcher.DEFAULT_INTERVAL;
+import static com.dke.data.agrirouter.impl.messaging.rest.MessageFetcher.MAX_TRIES_BEFORE_FAILURE;
+
+public class CloudOnboardingServiceImpl
+    implements CloudOnboardingService, MessageSender, ResponseValidator, MessageEncoder {
 
   private final EncodeMessageService encodeMessageService;
   private final FetchMessageService fetchMessageService;
   private final DecodeMessageService decodeMessageService;
 
-  public OnboardingServiceImpl() {
+  public CloudOnboardingServiceImpl() {
     this.encodeMessageService = new EncodeMessageServiceImpl();
     this.fetchMessageService = new FetchMessageServiceImpl();
     this.decodeMessageService = new DecodeMessageServiceImpl();
@@ -50,7 +49,7 @@ public class OnboardingServiceImpl
    * @return -
    */
   @Override
-  public List<OnboardingResponse> onboard(CloudOnboardingParameters parameters) {
+  public List<OnboardingResponse> send(CloudOnboardingParameters parameters) {
     parameters.validate();
     EncodedMessage encodedMessageResponse = this.encode(parameters);
     SendMessageParameters sendMessageParameters =
